Add files via upload
For a web development assignment, I created ‘Alpenliebe’, an engaging and visually appealing webpage dedicated to my favorite candy. This project involved an immersive layout with liquid design, ensuring a centered and responsive experience on different devices. I employed semantic HTML for structure and a separate CSS file for styling, including custom fonts, gradient backgrounds, and hover effects, to enhance the visual appeal. Additionally, the site featured a footer with a logo, company details, and a mock contact form. The page's design and functionalities were meticulously validated for both HTML and CSS standards, resulting in a professional and delightful user experience.